requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
pandas==2.2.1
pgvector==0.2.5
psycopg2-binary==2.9.9
Pygments==2.11.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==5.4.1
six==1.16.0
tzdata==2024.1
Werkzeug==3.0.1
